Module hashglobe::fake

This module contains shims around the stdlib HashMap that add fallible methods

These methods are a lie. They are not actually fallible. This is just to make it smooth to switch between hashmap impls in a codebase.

Structs

HashMap

HashSet

MapIter
An iterator over the entries of a HashMap.

MapIterMut
A mutable iterator over the entries of a HashMap.

RandomState
RandomState is the default state for HashMap types.

SetIntoIter
An owning iterator over the items of a HashSet.

SetIter
An iterator over the items of a HashSet.

Enums

Entry
A view into a single entry in a map, which may either be vacant or occupied.