Module hashglobe::hash_map

Structs

DefaultHasher
The default Hasher used by RandomState.

DefaultResizePolicy
The default behavior of HashMap implements a maximum load factor of 90.9%.

Drain
A draining iterator over the entries of a HashMap.

HashMap
A hash map implemented with linear probing and Robin Hood bucket stealing.

IntoIter
An owning iterator over the entries of a HashMap.

Iter
An iterator over the entries of a HashMap.

IterMut
A mutable iterator over the entries of a HashMap.

Keys
An iterator over the keys of a HashMap.

OccupiedEntry
A view into an occupied entry in a HashMap. It is part of the Entry enum.

RandomState
RandomState is the default state for HashMap types.

VacantEntry
A view into a vacant entry in a HashMap. It is part of the Entry enum.

Values
An iterator over the values of a HashMap.

ValuesMut
A mutable iterator over the values of a HashMap.

Enums

Entry
A view into a single entry in a map, which may either be vacant or occupied.

InternalEntry

VacantEntryState
Possible states of a VacantEntry.

Constants

DISPLACEMENT_THRESHOLD

MIN_NONZERO_RAW_CAPACITY

Functions

assert_covariance

pop_internal

robin_hood
Perform robin hood bucket stealing at the given bucket. You must also pass that bucket's displacement so we don't have to recalculate it.

search_hashed
Search for a pre-hashed key.