Struct hashglobe::table::Bucket

pub struct Bucket<K, V, M> {
    raw: RawBucket<K, V>,
    table: M,
}

Fields
raw: RawBucket<K, V>table: M
Implementations

impl<K, V, M> Bucket<K, V, M>

pub fn index(&self) -> usize

Get the raw index.

pub fn into_table(self) -> M

get the table.

impl<K, V, M: Deref<Target = RawTable<K, V>>> Bucket<K, V, M>

pub fn new(table: M, hash: SafeHash) -> Bucket<K, V, M>

pub fn new_from(r: RawBucket<K, V>, t: M) -> Bucket<K, V, M>

pub fn at_index(table: M, ib_index: usize) -> Bucket<K, V, M>

pub fn first(table: M) -> Bucket<K, V, M>

pub fn head_bucket(table: M) -> Bucket<K, V, M>

pub fn peek(self) -> BucketState<K, V, M>

Reads a bucket at a given index, returning an enum indicating whether it’s initialized or not. You need to match on this enum to get the appropriate types to call most of the other functions in this module.

pub fn next(&mut self)

Modifies the bucket in place to make it point to the next slot.

pub fn prev(&mut self)

Modifies the bucket in place to make it point to the previous slot.