Struct hashglobe::table::RawTable

pub struct RawTable<K, V> {
    capacity_mask: usize,
    size: usize,
    hashes: TaggedHashUintPtr,
    marker: PhantomData<(K, V)>,
}

The raw hashtable, providing safe-ish access to the unzipped and highly optimized arrays of hashes, and key-value pairs.

This design is a lot faster than the naive Vec<Option<(u64, K, V)>>, because we don’t pay for the overhead of an option on every element, and we get a generally more cache-aware design.

Essential invariants of this structure:

if t.hashes[i] == EMPTY_BUCKET, then Bucket::at_index(&t, i).raw points to ‘undefined’ contents. Don’t read from it. This invariant is enforced outside this module with the EmptyBucket, FullBucket, and SafeHash types.

An EmptyBucket is only constructed at an index with a hash of EMPTY_BUCKET.

A FullBucket is only constructed at an index with a non-EMPTY_BUCKET hash.

A SafeHash is only constructed for non-EMPTY_BUCKET hash. We get around hashes of zero by changing them to 0x8000_0000_0000_0000, which will likely map to the same bucket, while not being confused with “empty”.

Both “arrays represented by pointers” are the same length: capacity. This is set at creation and never changes. The arrays are unzipped and are more cache aware (scanning through 8 hashes brings in at most 2 cache lines, since they’re all right beside each other). This layout may waste space in padding such as in a map from u64 to u8, but is a more cache conscious layout as the key-value pairs are only very shortly probed and the desired value will be in the same or next cache line.

You can kind of think of this module/data structure as a safe wrapper around just the “table” part of the hashtable. It enforces some invariants at the type level and employs some performance trickery, but in general is just a tricked out Vec<Option<(u64, K, V)>>.

The hashtable also exposes a special boolean tag. The tag defaults to false when the RawTable is created and is accessible with the tag and set_tag functions.

Fields
capacity_mask: usizesize: usizehashes: TaggedHashUintPtrmarker: PhantomData<(K, V)>
Implementations

impl<K, V> RawTable<K, V>

unsafe fn new_uninitialized(capacity: usize) -> RawTable<K, V>

unsafe fn try_new_uninitialized(
capacity: usize
) -> Result<RawTable<K, V>, FailedAllocationError>

Does not initialize the buckets. The caller should ensure they, at the very least, set every hash to EMPTY_BUCKET.

fn raw_bucket_at(&self, index: usize) -> RawBucket<K, V>

pub fn new(capacity: usize) -> Result<RawTable<K, V>, FailedAllocationError>

Creates a new raw table from a given capacity. All buckets are initially empty.

pub fn capacity(&self) -> usize

The hashtable’s capacity, similar to a vector’s.

pub fn size(&self) -> usize

The number of elements ever put in the hashtable, minus the number of elements ever taken.

fn raw_buckets(&self) -> RawBuckets<'_, K, V>

pub fn iter(&self) -> Iter<'_, K, V>

pub fn iter_mut(&mut self) -> IterMut<'_, K, V>

pub fn into_iter(self) -> IntoIter<K, V>

pub fn drain(&mut self) -> Drain<'_, K, V>

unsafe fn rev_drop_buckets(&mut self)

Drops buckets in reverse order. It leaves the table in an inconsistent state and should only be used for dropping the table’s remaining entries. It’s used in the implementation of Drop.

pub fn set_tag(&mut self, value: bool)

Set the table tag

pub fn tag(&self) -> bool

Get the table tag