Trait hashglobe::table::Put

pub trait Put<K, V> {
    unsafe fn borrow_table_mut(&mut self) -> &mut RawTable<K, V>;
}

Put is implemented for types which provide access to a table and cannot be invalidated by filling a bucket. A similar implementation for Take is possible.

Required methods

unsafe fn borrow_table_mut(&mut self) -> &mut RawTable<K, V>

Loading content...