Struct hashglobe::table::FullBucket

pub struct FullBucket<K, V, M> {
    raw: RawBucket<K, V>,
    table: M,
}

Fields
raw: RawBucket<K, V>table: M
Implementations

impl<K, V, M> FullBucket<K, V, M>

pub fn table(&self) -> &M

Borrow a reference to the table.

pub fn table_mut(&mut self) -> &mut M

Borrow a mutable reference to the table.

pub fn into_table(self) -> M

Move out the reference to the table.

pub fn index(&self) -> usize

Get the raw index.

pub fn raw(&self) -> RawBucket<K, V>

Get the raw bucket.

impl<K, V, M: Deref<Target = RawTable<K, V>>> FullBucket<K, V, M>

pub fn next(self) -> Bucket<K, V, M>

pub fn into_bucket(self) -> Bucket<K, V, M>

pub fn stash(self) -> FullBucket<K, V, Self>

Duplicates the current position. This can be useful for operations on two or more buckets.

pub fn displacement(&self) -> usize

Get the distance between this bucket and the ‘ideal’ location as determined by the key’s hash stored in it.

In the cited blog posts above, this is called the “distance to initial bucket”, or DIB. Also known as “probe count”.

pub fn hash(&self) -> SafeHash

pub fn read(&self) -> (&K, &V)

Gets references to the key and value at a given index.

impl<'t, K, V> FullBucket<K, V, &'t mut RawTable<K, V>>

pub fn take(self) -> (EmptyBucket<K, V, &'t mut RawTable<K, V>>, K, V)

Removes this bucket’s key and value from the hashtable.

This works similarly to put, building an EmptyBucket out of the taken bucket.

impl<K, V, M> FullBucket<K, V, M> where
M: Put<K, V>,

pub fn replace(&mut self, h: SafeHash, k: K, v: V) -> (SafeHash, K, V)

impl<K, V, M> FullBucket<K, V, M> where
M: Deref<Target = RawTable<K, V>> + DerefMut,

pub fn read_mut(&mut self) -> (&mut K, &mut V)

Gets mutable references to the key and value at a given index.

impl<'t, K, V, M> FullBucket<K, V, M> where
M: Deref<Target = RawTable<K, V>> + 't,

pub fn into_refs(self) -> (&'t K, &'t V)

Exchange a bucket state for immutable references into the table. Because the underlying reference to the table is also consumed, no further changes to the structure of the table are possible; in exchange for this, the returned references have a longer lifetime than the references returned by read().

impl<'t, K, V, M> FullBucket<K, V, M> where
M: Deref<Target = RawTable<K, V>> + DerefMut + 't,

pub fn into_mut_refs(self) -> (&'t mut K, &'t mut V)

This works similarly to into_refs, exchanging a bucket state for mutable references into the table.